Add regex, <, >, != etc and the field type number
Issue: #8 (https://github.com/chovanecm/sacredboard/issues/8)
Because the database schema is not strictly defined,
we cannot assume a field in the dabase has a certain type.
Therefore, the user has to specify if he wants to treat the input as a number or as a string.

TODO: don't allow using 'number' type when the operator is 'regex' or when the value itself
is not a number.